+++ src/extension.js
@@ -5,7 +5,7 @@ import * as vscode from 'vscode';
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 
-const RpgleFreeX = require('./RpgleFree');
+import { RpgleFree as RpgleFreeX } from './RpgleFree.mjs';
 
 /**
  * This procedure converts the highlighted text to free form
@@ -40,7 +40,7 @@ function RpgleFree() {
 
   // Replace the text
   editor.edit(editBuilder => {
-    editBuilder.replace(curRange,lines.join(eol));
+    editBuilder.replace(curRange, lines.join(eol));
   })
 
   vscode.window.showInformationMessage('Selected text converted to free format');
@@ -51,20 +51,20 @@ function RpgleFree() {
 
 export function activate(context) {
 
-	// Use the console to output diagnostic information (console.log) and errors (console.error)
-	// This line of code will only be executed once when your extension is activated
-	console.log('Congratulations, your extension "vscode-rpgfree" is now active!');
+  // Use the console to output diagnostic information (console.log) and errors (console.error)
+  // This line of code will only be executed once when your extension is activated
+  console.log('Congratulations, your extension "vscode-rpgfree" is now active!');
 
-	// The command has been defined in the package.json file
-	// Now provide the implementation of the command with registerCommand
-	// The commandId parameter must match the command field in package.json
-	context.subscriptions.push(
-		vscode.commands.registerCommand('vscode-rpgfree.rpgleFree', function () {
-			RpgleFree();
-		})
-	);
+  // The command has been defined in the package.json file
+  // Now provide the implementation of the command with registerCommand
+  // The commandId parameter must match the command field in package.json
+  context.subscriptions.push(
+    vscode.commands.registerCommand('vscode-rpgfree.rpgleFree', function () {
+      RpgleFree();
+    })
+  );
 
 }
 
 // this method is called when your extension is deactivated
-export function deactivate() {}
+export function deactivate() { }
