--- src/specs/F.js
@@ -1,79 +1,0 @@
-module.exports = {
-  Parse: function (input, indent, wasSub) {
-    let output = {
-      remove: false,
-      change: false,
-      value: '',
-
-      beforeSpaces: 0,
-      nextSpaces: 0
-    };
-
-    let name = input.substring(7, 17).trim(); //File name
-    let type = input.substring(17, 18).toUpperCase(); // I, U, O, C
-    let fileadd = input.substring(20, 21).toUpperCase(); // A
-    let external = input.substring(22, 23).toUpperCase(); // F, E
-    let recordLength = input.substring(23, 28).toUpperCase();
-    let field = input.substring(34, 35).toUpperCase(); //KEYED
-    let device = input.substring(36, 43).toUpperCase().trim(); //device: DISK, WORKSTN
-    let keywords = input.substring(44).trim();
-
-    output.value = `Dcl-F ${name}`;
-
-    switch (type) {
-      case 'I':
-        if (fileadd == 'A')
-          type = '*Input:*Output';
-        else
-          type = '*Input';
-        break;
-
-      case 'U':
-        type = '*Update:*Delete:*Output';
-        break;
-
-      case 'O':
-        if (device != 'PRINTER')
-          type = '*Output';
-        else
-          type = '';
-        break;
-
-      case 'C':
-        if (device != 'WORKSTN')
-          type = '*INPUT:*OUTPUT';
-        else
-          type = '';
-        break;
-
-      default:
-        type = '';
-        break;
-    }
-
-    if (external != 'E') {
-      device = `${device}(${recordLength.trim()})`;
-    }
-    if (device != 'DISK') {
-      output.value += ` ${device}`;
-    }
-    if (type != '') {
-      output.value += ` Usage(${type})`;
-    }
-    if (field == 'K') {
-      output.value += ' Keyed';
-    }
-    if (keywords != '') {
-      if (name == '') {
-        output.aboveKeywords = keywords;
-      } else {
-        output.value += ` ${keywords}`;
-      }
-    }
-    if (output.value !== '') {
-      output.change = true;
-      output.value = `${output.value.trimRight()};`;
-    }
-    return output;
-  }
-}
